backend/requirements.txt
 Flask
 flask-cors
 mysql-connector-python
Werkzeug
bcrypt
Werkzeug
mysql-connector-python
requests